Remove duplicate Goal class
