+++ build.gradle
@@ -25,7 +25,8 @@ repositories {
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    compile group: 'io.jsonwebtoken', name: 'jjwt', version: '0.7.0'
+//    compile group: 'io.jsonwebtoken', name: 'jjwt', version: '0.7.0'
+    implementation 'io.jsonwebtoken:jjwt:0.2'
     implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.security:spring-security-test'
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
