+++ src/main/java/com/goal/restservice/domain/Goal.java
@@ -1,11 +1,17 @@
 package com.goal.restservice.domain;
 
+import java.time.LocalDate;
+import javax.persistence.FetchType;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import lombok.*;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
 
 @Entity
 @Builder
@@ -13,12 +19,21 @@ import javax.persistence.Id;
 @AllArgsConstructor
 @Getter
 @Setter
-public class Goal {
+public class Goal extends BaseTimeEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  Long id;
+  private Long id;
 
-  String category;
-  String desc;
+  @ManyToOne
+  @JoinColumn
+  private Category category;
+
+  @ManyToOne(fetch = FetchType.EAGER, optional = false)
+  @OnDelete(action = OnDeleteAction.CASCADE)
+  private User user;
+
+  private String title;
+  private String desc;
+  private LocalDate dueDate;
 }
