--- src/main/java/com/goal/restservice/domain/goals/Goal.java
@@ -1,42 +1,0 @@
-package com.goal.restservice.domain.goals;
-
-import com.goal.restservice.domain.BaseTimeEntity;
-import com.goal.restservice.domain.User;
-import com.goal.restservice.domain.categories.Category;
-import java.time.LocalDate;
-import javax.persistence.FetchType;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import lombok.*;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
-@Entity
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class Goal extends BaseTimeEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @ManyToOne
-  @JoinColumn
-  private Category category;
-
-  @ManyToOne(fetch = FetchType.EAGER, optional = false)
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  private User user;
-
-  private String title;
-  private String desc;
-  private LocalDate dueDate;
-}
