+++ src/main/java/com/goal/restservice/repository/GoalRepository.java
@@ -1,8 +1,7 @@
 package com.goal.restservice.repository;
 
-import com.goal.restservice.domain.goals.Goal;
+import com.goal.restservice.domain.Goal;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GoalRepository extends JpaRepository<Goal, Long> {
