+++ src/main/java/com/goal/restservice/service/CategoryService.java
@@ -1,10 +1,7 @@
 package com.goal.restservice.service;
 
-import com.goal.restservice.domain.categories.Category;
-import com.goal.restservice.repository.CategoryRepository;
+import com.goal.restservice.domain.Category;
 import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 public interface CategoryService {
