+++ src/main/java/com/goal/restservice/service/CategoryServiceImpl.java
@@ -1,9 +1,8 @@
 package com.goal.restservice.service;
 
-import com.goal.restservice.domain.categories.Category;
+import com.goal.restservice.domain.Category;
 import com.goal.restservice.repository.CategoryRepository;
 import java.util.List;
-import java.util.Optional;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
